jump_force/functions/jump_height/pyproject.toml
[tool.poetry]
name = "application_analyze"
version = "0.1.0"
description = "analyze application"
authors = ["Joshua Schmidt <jns223@cornell.edu>"]
license = "CC-BY-NC-ND-4.0"

[tool.poetry.dependencies]
python = "^3.9"
loguru = "^0.6.0"
opencv-python = "^4.5.5"
awslambdaric = "^2.0.1"
boto3 = "^1.21.7"
python-dotenv = "^0.19.2"
boto3-type-annotations = "^0.3.1"
requests-aws4auth = "^1.1.1"
requests-toolbelt = "^0.9.1"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"
mypy = "^0.931"
pylint = "^2.12.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
